Make unsigned into uint64_t and int into int64_t on axi4mlir lib

+++ mlir/include/mlir/ExecutionEngine/axi/accelerators/mm_4x4_v1/dma_engine.sc.h
@@ -76,8 +76,8 @@ SC_MODULE(DMA_DRIVER) {
     reset_signal_is(reset, true);
   }
 
-  int *DMA_input_buffer;
-  int *DMA_output_buffer;
+  int64_t *DMA_input_buffer;
+  int64_t *DMA_output_buffer;
 
   // TODO: input_length = Number of elements * (sizeof(elements)/32)
   int input_len;
