+++ mlir/include/mlir/ExecutionEngine/axi/api_v1.h
@@ -32,7 +32,7 @@
 // Simple view of DMA
 /*
 dma -> {
-    control_register_address : unsigned int     # Mapped address to the start of
+    control_register_address : uint64_t     # Mapped address to the start of
 the DMA control registers Buffer input_buffer (address,size) Buffer
 output_buffer  (address,size)
 }
@@ -68,21 +68,21 @@ struct dma {
 #define LOG(x)
 #endif
 
-  unsigned int *dma_address;
-  unsigned int *dma_input_address;
-  unsigned int *dma_output_address;
-  unsigned int dma_input_buffer_size;
-  unsigned int dma_output_buffer_size;
-  unsigned int dma_input_paddress;
-  unsigned int dma_output_paddress;
-  unsigned int *acc_address;
-  unsigned int current_input_offset;
+  uint64_t *dma_address;
+  uint64_t *dma_input_address;
+  uint64_t *dma_output_address;
+  uint64_t dma_input_buffer_size;
+  uint64_t dma_output_buffer_size;
+  uint64_t dma_input_paddress;
+  uint64_t dma_output_paddress;
+  uint64_t *acc_address;
+  uint64_t current_input_offset;
 
   // Profiling Variables
-  unsigned int dma_send_length=0;
-  unsigned int dma_recv_length=0;
-  unsigned int dma_send_count=0;
-  unsigned int dma_recv_count=0;
+  uint64_t dma_send_length=0;
+  uint64_t dma_recv_length=0;
+  uint64_t dma_send_count=0;
+  uint64_t dma_recv_count=0;
 
 
   // temp --- need to remove later
@@ -93,10 +93,10 @@ struct dma {
   DMA_DRIVER *dmad;
 #endif
 
-  void dma_init(unsigned int dma_address, unsigned int dma_input_address,
-                unsigned int dma_input_buffer_size,
-                unsigned int dma_output_address,
-                unsigned int dma_output_buffer_size);
+  void dma_init(uint64_t dma_address, uint64_t dma_input_address,
+                uint64_t dma_input_buffer_size,
+                uint64_t dma_output_address,
+                uint64_t dma_output_buffer_size);
 
   // Memory unmaps DMA control_register_address and Input and output buffers
   void dma_free();
@@ -107,34 +107,34 @@ struct dma {
   //-----------------BUFFER Functions-----------------
   // Get the MMap address of the input buffer of the dma  *Needed to copy data
   // to Input_Buffer*
-  unsigned int *dma_get_inbuffer();
+  uint64_t *dma_get_inbuffer();
 
   // Get the MMap address of the output buffer of the dma *Needed to copy data
   // from Output_Buffer*
-  unsigned int *dma_get_outbuffer();
+  uint64_t *dma_get_outbuffer();
 
   //================================================================================================================
 
   //-----------------BUFFER Functions-----------------
   // Copy data into the Input Buffer (length to write, offset to write to)
   // returns 0 if successful
-  int dma_copy_to_inbuffer(unsigned int *host_src_address, int data_length,
-                           int offset);
+  int64_t dma_copy_to_inbuffer(uint64_t *host_src_address, int64_t data_length,
+                           int64_t offset);
 
   template <typename T>
-  int mlir_dma_copy_to_inbuffer(T *mr_base, int64_t mr_dim, int64_t mr_rank,
+  int64_t mlir_dma_copy_to_inbuffer(T *mr_base, int64_t mr_dim, int64_t mr_rank,
                                 int64_t mr_offset, const int64_t *mr_sizes,
-                                const int64_t *mr_strides, int dma_offset);
+                                const int64_t *mr_strides, int64_t dma_offset);
 
   // Copy data from the Output Buffer (length to read, offset to read from)
   // returns 0 if successful
-  int dma_copy_from_outbuffer(unsigned int *host_dst_address, int data_length,
-                              int offset);
+  int64_t dma_copy_from_outbuffer(uint64_t *host_dst_address, int64_t data_length,
+                              int64_t offset);
 
   template <typename T>
-  int mlir_dma_copy_from_outbuffer(T *mr_base, int64_t mr_dim, int64_t mr_rank,
+  int64_t mlir_dma_copy_from_outbuffer(T *mr_base, int64_t mr_dim, int64_t mr_rank,
                                    int64_t mr_offset, const int64_t *mr_sizes,
-                                   const int64_t *mr_strides, int dma_offset);
+                                   const int64_t *mr_strides, int64_t dma_offset);
 
   //============================================================================
 
@@ -145,13 +145,13 @@ struct dma {
    * Starts transfers to the accelerator using dma associated with dma_id
    * Return 0 if successful, returns negative if error occurs
    */
-  int dma_start_send(int length, int offset);
+  int64_t dma_start_send(int64_t length, int64_t offset);
 
   // Blocks thread until dma MMS2 transfer is complete
   void dma_wait_send();
 
   // Same as dma_send but thread does not block, returns 0 if done
-  int dma_check_send();
+  int64_t dma_check_send();
 
   //-----------------DMA S2MM Functions-----------------
   /**
@@ -160,24 +160,24 @@ struct dma {
    * Starts storing data recieved through dma associated with dma_id
    * Return 0 if successful, returns negative if error occurs
    */
-  int dma_start_recv(int length, int offset);
+  int64_t dma_start_recv(int64_t length, int64_t offset);
 
   // Blocks thread until dma S2MM transfer is complete (TLAST signal is seen)
   void dma_wait_recv();
 
   // Same as dma_recv but thread does not block, returns 0 if done
-  int dma_check_recv();
+  int64_t dma_check_recv();
 
   //********************************** Unexposed Functions
   //**********************************
   void initDMAControls();
-  void dma_set(unsigned int *dma_virtual_address, int offset,
-               unsigned int value);
-  unsigned int dma_get(unsigned int *dma_virtual_address, int offset);
+  void dma_set(uint64_t *dma_virtual_address, int64_t offset,
+               uint64_t value);
+  uint64_t dma_get(uint64_t *dma_virtual_address, int64_t offset);
   void dma_mm2s_sync();
   void dma_s2mm_sync();
-  void acc_init(unsigned int base_addr, int length);
-  void dump_acc_signals(int state);
+  void acc_init(uint64_t base_addr, int64_t length);
+  void dump_acc_signals(int64_t state);
 };
 
 #endif
