+++ mlir/lib/ExecutionEngine/axi/api_v1.cpp
@@ -12,10 +12,10 @@
 
 #include "mlir/ExecutionEngine/axi/api_v1.h"
 
-void dma::dma_init(unsigned int _dma_address, unsigned int _dma_input_address,
-                   unsigned int _dma_input_buffer_size,
-                   unsigned int _dma_output_address,
-                   unsigned int _dma_output_buffer_size) {
+void dma::dma_init(uint64_t _dma_address, uint64_t _dma_input_address,
+                   uint64_t _dma_input_buffer_size,
+                   uint64_t _dma_output_address,
+                   uint64_t _dma_output_buffer_size) {
   int dh = open("/dev/mem", O_RDWR | O_SYNC);
   void *dma_mm = mmap(NULL, PAGE_SIZE, PROT_READ | PROT_WRITE, MAP_SHARED, dh,
                       _dma_address); // Memory map AXI Lite register block
@@ -25,9 +25,9 @@ void dma::dma_init(unsigned int _dma_address, unsigned int _dma_input_address,
   void *dma_out_mm =
       mmap(NULL, _dma_output_buffer_size, PROT_READ, MAP_SHARED, dh,
            _dma_output_address); // Memory map destination address
-  dma_address = reinterpret_cast<unsigned int *>(dma_mm);
-  dma_input_address = reinterpret_cast<unsigned int *>(dma_in_mm);
-  dma_output_address = reinterpret_cast<unsigned int *>(dma_out_mm);
+  dma_address = reinterpret_cast<uint64_t *>(dma_mm);
+  dma_input_address = reinterpret_cast<uint64_t *>(dma_in_mm);
+  dma_output_address = reinterpret_cast<uint64_t *>(dma_out_mm);
   dma_input_buffer_size = _dma_input_buffer_size;
   dma_output_buffer_size = _dma_output_buffer_size;
   dma_input_paddress = _dma_input_address;
@@ -46,23 +46,23 @@ void dma::dma_free() {
 
 // We could reduce to one set of the following calls
 //==============================================================================
-unsigned int *dma::dma_get_inbuffer() { return dma_input_address; }
+uint64_t *dma::dma_get_inbuffer() { return dma_input_address; }
 
-unsigned int *dma::dma_get_outbuffer() { return dma_output_address; }
+uint64_t *dma::dma_get_outbuffer() { return dma_output_address; }
 //==============================================================================
-int dma::dma_copy_to_inbuffer(unsigned int *src_address, int data_length,
-                              int offset) {
+int64_t dma::dma_copy_to_inbuffer(uint64_t *src_address, int64_t data_length,
+                              int64_t offset) {
   m_assert("data copy will overflow input buffer",
-           (unsigned int)(offset + data_length) <= dma_input_buffer_size);
+           (uint64_t)(offset + data_length) <= dma_input_buffer_size);
   std::memcpy(dma_input_address + offset, src_address, data_length * 4);
   current_input_offset += data_length;
   return 0;
 }
 
-int dma::dma_copy_from_outbuffer(unsigned int *dst_address, int data_length,
-                                 int offset) {
+int64_t dma::dma_copy_from_outbuffer(uint64_t *dst_address, int64_t data_length,
+                                 int64_t offset) {
   m_assert("tries to access data outwith the output buffer",
-           (unsigned int)(offset + data_length) <= dma_output_buffer_size);
+           (uint64_t)(offset + data_length) <= dma_output_buffer_size);
   std::memcpy(dst_address, dma_output_address + offset, data_length * 4);
   return 0;
 }
@@ -72,7 +72,7 @@ template <typename T>
 inline void copy_memref_to_array(T *mr_base, int64_t mr_dim, int64_t mr_rank,
                                  int64_t mr_offset, const int64_t *mr_sizes,
                                  const int64_t *mr_strides,
-                                 unsigned int *dst_base,
+                                 uint64_t *dst_base,
                                  const int64_t dst_offset) {
   int64_t rank = mr_rank;
   // Handle empty shapes -> nothing to copy.
@@ -84,7 +84,7 @@ inline void copy_memref_to_array(T *mr_base, int64_t mr_dim, int64_t mr_rank,
   srcPtr = mr_base + mr_offset;
 
   // TODO make templated
-  unsigned int *dstPtr;
+  uint64_t *dstPtr;
   dstPtr = dst_base + dst_offset;
 
   if (rank == 0) {
@@ -141,9 +141,9 @@ inline void copy_memref_to_array(T *mr_base, int64_t mr_dim, int64_t mr_rank,
 
 // Implements the actual copy
 template <typename T>
-int dma::mlir_dma_copy_to_inbuffer(T *mr_base, int64_t mr_dim, int64_t mr_rank,
+int64_t dma::mlir_dma_copy_to_inbuffer(T *mr_base, int64_t mr_dim, int64_t mr_rank,
                                    int64_t mr_offset, const int64_t *mr_sizes,
-                                   const int64_t *mr_strides, int dma_offset) {
+                                   const int64_t *mr_strides, int64_t dma_offset) {
   std::cout << __FILE__ << ": " << __LINE__ << " [" << __func__ << "]\n";
 
   copy_memref_to_array(mr_base, mr_dim, mr_rank, mr_offset, mr_sizes,
@@ -157,7 +157,7 @@ template <typename T>
 inline void copy_array_to_memref(T *mr_base, int64_t mr_dim, int64_t mr_rank,
                                  int64_t mr_offset, const int64_t *mr_sizes,
                                  const int64_t *mr_strides,
-                                 unsigned int *src_base,
+                                 uint64_t *src_base,
                                  const int64_t src_offset) {
   int64_t rank = mr_rank;
   // Handle empty shapes -> nothing to copy.
@@ -169,7 +169,7 @@ inline void copy_array_to_memref(T *mr_base, int64_t mr_dim, int64_t mr_rank,
   dstPtr = mr_base + mr_offset;
 
   // TODO make templated
-  unsigned int *srcPtr;
+  uint64_t *srcPtr;
   srcPtr = src_base + src_offset;
 
   if (rank == 0) {
@@ -226,11 +226,11 @@ inline void copy_array_to_memref(T *mr_base, int64_t mr_dim, int64_t mr_rank,
 }
 
 template <typename T>
-int dma::mlir_dma_copy_from_outbuffer(T *mr_base, int64_t mr_dim,
+int64_t dma::mlir_dma_copy_from_outbuffer(T *mr_base, int64_t mr_dim,
                                       int64_t mr_rank, int64_t mr_offset,
                                       const int64_t *mr_sizes,
                                       const int64_t *mr_strides,
-                                      int dma_offset) {
+                                      int64_t dma_offset) {
 
   std::cout << __FILE__ << ": " << __LINE__ << " [" << __func__ << "]\n";
 
@@ -241,25 +241,25 @@ int dma::mlir_dma_copy_from_outbuffer(T *mr_base, int64_t mr_dim,
 }
 
 // Make templates concrete:
-template int dma::mlir_dma_copy_to_inbuffer<float>(
+template int64_t dma::mlir_dma_copy_to_inbuffer<float>(
     float *mr_base, int64_t mr_dim, int64_t mr_rank, int64_t mr_offset,
-    const int64_t *mr_sizes, const int64_t *mr_strides, int dma_offset);
+    const int64_t *mr_sizes, const int64_t *mr_strides, int64_t dma_offset);
 
-template int dma::mlir_dma_copy_to_inbuffer<int>(
-    int *mr_base, int64_t mr_dim, int64_t mr_rank, int64_t mr_offset,
-    const int64_t *mr_sizes, const int64_t *mr_strides, int dma_offset);
+template int64_t dma::mlir_dma_copy_to_inbuffer<int64_t>(
+    int64_t *mr_base, int64_t mr_dim, int64_t mr_rank, int64_t mr_offset,
+    const int64_t *mr_sizes, const int64_t *mr_strides, int64_t dma_offset);
 
-template int dma::mlir_dma_copy_from_outbuffer<float>(
+template int64_t dma::mlir_dma_copy_from_outbuffer<float>(
     float *mr_base, int64_t mr_dim, int64_t mr_rank, int64_t mr_offset,
-    const int64_t *mr_sizes, const int64_t *mr_strides, int dma_offset);
+    const int64_t *mr_sizes, const int64_t *mr_strides, int64_t dma_offset);
 
-template int dma::mlir_dma_copy_from_outbuffer<int>(
-    int *mr_base, int64_t mr_dim, int64_t mr_rank, int64_t mr_offset,
-    const int64_t *mr_sizes, const int64_t *mr_strides, int dma_offset);
+template int64_t dma::mlir_dma_copy_from_outbuffer<int64_t>(
+    int64_t *mr_base, int64_t mr_dim, int64_t mr_rank, int64_t mr_offset,
+    const int64_t *mr_sizes, const int64_t *mr_strides, int64_t dma_offset);
 
-int dma::dma_start_send(int length, int offset) {
+int64_t dma::dma_start_send(int64_t length, int64_t offset) {
   m_assert("trying to send data outside the input buffer",
-           (unsigned int)(offset + length) <= dma_input_buffer_size);
+           (uint64_t)(offset + length) <= dma_input_buffer_size);
   dma_set(dma_address, MM2S_START_ADDRESS, dma_input_paddress + (offset * 4));
   msync(dma_address, PAGE_SIZE, MS_SYNC);
   dma_set(dma_address, MM2S_LENGTH, length * 4);
@@ -274,8 +274,8 @@ void dma::dma_wait_send() {
   LOG("Data Transfer - Done");
 }
 
-int dma::dma_check_send() {
-  unsigned int mm2s_status = dma_get(dma_address, MM2S_STATUS_REGISTER);
+int64_t dma::dma_check_send() {
+  uint64_t mm2s_status = dma_get(dma_address, MM2S_STATUS_REGISTER);
   bool done = !((!(mm2s_status & 1 << 12)) || (!(mm2s_status & 1 << 1)));
   if (done)
     LOG("Data Transfer - Done");
@@ -284,9 +284,9 @@ int dma::dma_check_send() {
   return done ? 0 : -1;
 }
 
-int dma::dma_start_recv(int length, int offset) {
+int64_t dma::dma_start_recv(int64_t length, int64_t offset) {
   m_assert("trying receive data outside the output buffer",
-           (unsigned int)(offset + length) <= dma_output_buffer_size);
+           (uint64_t)(offset + length) <= dma_output_buffer_size);
   dma_set(dma_address, S2MM_DESTINATION_ADDRESS,
           dma_output_paddress + (offset * 4));
   msync(dma_address, PAGE_SIZE, MS_SYNC);
@@ -300,12 +300,12 @@ void dma::dma_wait_recv() {
   LOG("Data Receive - Waiting");
   LOG("Data Receive - Waiting " << dma_get(dma_address, S2MM_LENGTH));
   dma_s2mm_sync();
-  // unsigned int recv_len =  dma_get(dma_address,S2MM_LENGTH);
+  // uint64_t recv_len =  dma_get(dma_address,S2MM_LENGTH);
   LOG("Data Receive - Done " << dma_get(dma_address, S2MM_LENGTH));
 }
 
-int dma::dma_check_recv() {
-  unsigned int s2mm_status = dma_get(dma_address, S2MM_STATUS_REGISTER);
+int64_t dma::dma_check_recv() {
+  uint64_t s2mm_status = dma_get(dma_address, S2MM_STATUS_REGISTER);
   bool done = !((!(s2mm_status & 1 << 12)) || (!(s2mm_status & 1 << 1)));
   if (done)
     LOG("Data Receive - Done");
@@ -333,17 +333,17 @@ void dma::initDMAControls() {
   dma_set(dma_address, MM2S_CONTROL_REGISTER, 0xf001);
 }
 
-void dma::dma_set(unsigned int *dma_address, int offset, unsigned int value) {
+void dma::dma_set(uint64_t *dma_address, int64_t offset, uint64_t value) {
   dma_address[offset >> 2] = value;
 }
 
-unsigned int dma::dma_get(unsigned int *dma_address, int offset) {
+uint64_t dma::dma_get(uint64_t *dma_address, int64_t offset) {
   return dma_address[offset >> 2];
 }
 
 void dma::dma_mm2s_sync() {
   msync(dma_address, PAGE_SIZE, MS_SYNC);
-  unsigned int mm2s_status = dma_get(dma_address, MM2S_STATUS_REGISTER);
+  uint64_t mm2s_status = dma_get(dma_address, MM2S_STATUS_REGISTER);
   while (!(mm2s_status & 1 << 12) || !(mm2s_status & 1 << 1)) {
     msync(dma_address, PAGE_SIZE, MS_SYNC);
     mm2s_status = dma_get(dma_address, MM2S_STATUS_REGISTER);
@@ -352,15 +352,15 @@ void dma::dma_mm2s_sync() {
 
 void dma::dma_s2mm_sync() {
   msync(dma_address, PAGE_SIZE, MS_SYNC);
-  unsigned int s2mm_status = dma_get(dma_address, S2MM_STATUS_REGISTER);
+  uint64_t s2mm_status = dma_get(dma_address, S2MM_STATUS_REGISTER);
   while (!(s2mm_status & 1 << 12) || !(s2mm_status & 1 << 1)) {
     msync(dma_address, PAGE_SIZE, MS_SYNC);
     s2mm_status = dma_get(dma_address, S2MM_STATUS_REGISTER);
   }
 }
 
-void dma::acc_init(unsigned int base_addr, int length) {
-  int dh = open("/dev/mem", O_RDWR | O_SYNC);
+void dma::acc_init(uint64_t base_addr, int64_t length) {
+  int64_t dh = open("/dev/mem", O_RDWR | O_SYNC);
   size_t virt_base = base_addr & ~(PAGE_SIZE - 1);
   size_t virt_offset = base_addr - virt_base;
   void *addr = mmap(NULL, length + virt_offset, PROT_READ | PROT_WRITE,
@@ -368,10 +368,10 @@ void dma::acc_init(unsigned int base_addr, int length) {
   close(dh);
   if (addr == (void *)-1)
     exit(EXIT_FAILURE);
-  acc_address = reinterpret_cast<unsigned int *>(addr);
+  acc_address = reinterpret_cast<uint64_t *>(addr);
 }
 
-void dma::dump_acc_signals(int state) {
+void dma::dump_acc_signals(int64_t state) {
   msync(acc_address, PAGE_SIZE, MS_SYNC);
   std::ofstream file;
   file.open("dump_acc_signals.dat", std::ios_base::app);
